+++ src/Mailer/AbstractMailType.php
@@ -6,39 +6,15 @@ namespace Schvoy\MailTemplateBundle\Mailer;
 
 abstract class AbstractMailType implements MailTypeInterface
 {
-    protected string $subject;
+    protected string $translationKeyPath;
 
-    protected string $body;
-
-    protected array $configuration = [];
-
-    public function getSubject(): string
+    public function getTranslationKeyPath(): string
     {
-        return $this->subject;
+        return $this->translationKeyPath;
     }
 
-    public function setSubject(string $subject): void
+    public function setTranslationKeyPath(string $translationKeyPath): void
     {
-        $this->subject = $subject;
-    }
-
-    public function getBody(): string
-    {
-        return $this->body;
-    }
-
-    public function setBody(string $body): void
-    {
-        $this->body = $body;
-    }
-
-    public function getConfiguration(): array
-    {
-        return $this->configuration;
-    }
-
-    public function setConfiguration(array $configuration): void
-    {
-        $this->configuration = $configuration;
+        $this->translationKeyPath = $translationKeyPath;
     }
 }