+++ src/Mailer/MailTypeInterface.php
@@ -6,17 +6,9 @@ namespace Schvoy\MailTemplateBundle\Mailer;
 
 interface MailTypeInterface
 {
-    public function getSubject(): string;
+    public function getTranslationKeyPath(): string;
 
-    public function setSubject(string $subject): void;
+    public function setTranslationKeyPath(string $translationKeyPath): void;
 
-    public function getBody(): string;
-
-    public function setBody(string $body): void;
-
-    public function getContent(array $configuration): string;
-
-    public function getConfiguration(): array;
-
-    public function setConfiguration(array $configuration): void;
+    public function getContent(Configuration $configuration): string;
 }