+++ src/Mailer/Recipient.php
@@ -6,8 +6,12 @@ namespace Schvoy\MailTemplateBundle\Mailer;
 
 class Recipient
 {
-    public function __construct(private string $email, private ?string $name = null)
-    {
+    public function __construct(
+        private readonly string $email,
+        private readonly ?string $name = null,
+        private readonly bool $cc = false,
+        private readonly bool $bcc = false,
+    ) {
     }
 
     public function getName(): ?string
@@ -19,4 +23,14 @@ class Recipient
     {
         return $this->email;
     }
-}
+
+    public function isCc(): bool
+    {
+        return $this->cc;
+    }
+
+    public function isBcc(): bool
+    {
+        return $this->bcc;
+    }
+    }